refactor(report): 도메인 root flat 구조를 sub-folder로 분류
- root 32개 파일 → 5개(페이지)로 정리
- sections/ (8파일): Plan/Intro/Example/Review/Faq, ServiceProcess, Promo, ResearchTeam
- modal/ (1파일 추가): ReportApplyBottomSheet
- ui/header/ (4파일): Header, MainHeader, SubHeader, SectionHeader
- ui/heading/ (3파일): Heading1, Heading2, Label
- ui/card/ (3파일): ProgramCard, ReportInterviewCard, ReportExampleCard
- ui/ (8파일 평탄): Tooltip, ReportLayout, ReportNavigation, ReportProgramRecommendSlider, ReportContentContainer, ReportBasicInfo, DateTimePicker, FilledInput
- 도메인 내부/외부 import 절대경로(@/domain/report/...) 통일

거대 파일 4개(764/661/590/510줄)의 본체 분리는 별도 PR에서 진행.
