+++ apps/web/src/app/(user)/report/order/fail/page.tsx
@@ -4,9 +4,9 @@ import { convertReportPriceType, useGetReportDetailQuery } from '@/api/report';
 import { paymentFailSearchParamsSchema } from '@/data/getPaymentSearchParams';
 import ReportCreditRow from '@/domain/mypage/credit/ReportCreditRow';
 import ReportCreditSubRow from '@/domain/mypage/credit/ReportCreditSubRow';
-import Heading1 from '@/domain/report/Heading1';
-import Heading2 from '@/domain/report/Heading2';
-import Card from '@/domain/report/ProgramCard';
+import Heading1 from '@/domain/report/ui/heading/Heading1';
+import Heading2 from '@/domain/report/ui/heading/Heading2';
+import Card from '@/domain/report/ui/card/ProgramCard';
 import useReportPayment from '@/hooks/useReportPayment';
 import useReportProgramInfo from '@/hooks/useReportProgramInfo';
 import useReportApplicationStore from '@/store/useReportApplicationStore';
