+++ apps/web/src/app/(user)/report/order/result/page.tsx
@@ -10,9 +10,9 @@ import ReportCreditRow from '@/domain/mypage/credit/ReportCreditRow';
 import ReportCreditSubRow from '@/domain/mypage/credit/ReportCreditSubRow';
 import DescriptionBox from '@/domain/program/paymentSuccess/DescriptionBox';
 import PaymentInfoRow from '@/domain/program/paymentSuccess/PaymentInfoRow';
-import Heading1 from '@/domain/report/Heading1';
-import Heading2 from '@/domain/report/Heading2';
-import Card from '@/domain/report/ProgramCard';
+import Heading1 from '@/domain/report/ui/heading/Heading1';
+import Heading2 from '@/domain/report/ui/heading/Heading2';
+import Card from '@/domain/report/ui/card/ProgramCard';
 import useReportPayment from '@/hooks/useReportPayment';
 import useReportProgramInfo from '@/hooks/useReportProgramInfo';
 import dayjs from '@/lib/dayjs';
