+++ apps/web/src/app/(user)/report/payment/[reportType]/[reportId]/page.tsx
@@ -10,9 +10,9 @@ import BaseButton from '@/common/button/BaseButton';
 import BackHeader from '@/common/header/BackHeader';
 import Input from '@/common/input/v2/Input';
 import BottomSheet from '@/common/sheet/BottomSheeet';
-import Heading2 from '@/domain/report/Heading2';
-import Label from '@/domain/report/Label';
-import ProgramCard from '@/domain/report/ProgramCard';
+import Heading2 from '@/domain/report/ui/heading/Heading2';
+import Label from '@/domain/report/ui/heading/Label';
+import ProgramCard from '@/domain/report/ui/card/ProgramCard';
 import useReportPayment from '@/hooks/useReportPayment';
 import useReportProgramInfo from '@/hooks/useReportProgramInfo';
 import { generateOrderId } from '@/lib/order';
