+++ apps/web/src/domain/report/ReportApplyPage.tsx
@@ -15,15 +15,15 @@ import BaseButton from '@/common/button/BaseButton';
 import { OptionFormRadioControlLabel } from '@/common/ControlLabel';
 import BackHeader from '@/common/header/BackHeader';
 import HorizontalRule from '@/common/HorizontalRule';
-import RequiredStar from './ui/RequiredStar';
+import RequiredStar from '@/domain/report/ui/RequiredStar';
 import BottomSheet from '@/common/sheet/BottomSheeet';
 import { getReportThumbnail } from '@/domain/mypage/credit/CreditListItem';
-import DateTimePicker from '@/domain/report/DateTimePicker';
-import FilledInput from '@/domain/report/FilledInput';
-import Heading2 from '@/domain/report/Heading2';
-import Label from '@/domain/report/Label';
-import ProgramCard from '@/domain/report/ProgramCard';
-import Tooltip from '@/domain/report/Tooltip';
+import DateTimePicker from '@/domain/report/ui/DateTimePicker';
+import FilledInput from '@/domain/report/ui/FilledInput';
+import Heading2 from '@/domain/report/ui/heading/Heading2';
+import Label from '@/domain/report/ui/heading/Label';
+import ProgramCard from '@/domain/report/ui/card/ProgramCard';
+import Tooltip from '@/domain/report/ui/Tooltip';
 import useMinDate from '@/hooks/useMinDate';
 import useReportProgramInfo from '@/hooks/useReportProgramInfo';
 import useValidateUrl from '@/hooks/useValidateUrl';
