+++ apps/web/src/domain/report/ReportManagementPage.tsx
@@ -22,7 +22,7 @@ import {
   useGetMyReports,
 } from '@/api/report';
 import Badge from '@/common/badge/Badge';
-import Tooltip from '@/domain/report/Tooltip';
+import Tooltip from '@/domain/report/ui/Tooltip';
 import useActiveReports from '@/hooks/useActiveReports';
 import { download } from '@/lib/download';
 import { twMerge } from '@/lib/twMerge';
