+++ apps/web/src/domain/report/sections/ReportIntroSection.tsx
@@ -3,7 +3,7 @@ import Profile1 from '@/assets/illust/report_profile_1.svg?react';
 import Profile2 from '@/assets/illust/report_profile_2.svg?react';
 import Profile3 from '@/assets/illust/report_profile_3.svg?react';
 import { REPORT_INTRO } from '@/data/reportConstant';
-import { REPORT_INTRO_ID } from '@/domain/report/ReportNavigation';
+import { REPORT_INTRO_ID } from '@/domain/report/ui/ReportNavigation';
 import { personalStatementColors } from '@/domain/report/ReportPersonalStatementPage';
 import { resumeColors } from '@/domain/report/ReportResumePage';
 import { useMediaQuery } from '@mui/material';
