+++ apps/web/src/domain/report/ui/ReportBasicInfo.tsx
