+++ apps/web/src/domain/report/ui/ReportLayout.tsx
