+++ apps/web/src/domain/report/ui/ReportNavigation.tsx
