+++ apps/web/src/domain/report/ui/card/ReportExampleCard.tsx
