+++ apps/web/src/domain/report/ui/header/SectionHeader.tsx
